Fix SignalAction once and for all.
We don't need to wrap the lambda in std::function at all, and the problem of
being unable to convert it to a regular void(*)(int) was due to it using
captures by reference. There's no need for that at the current state (as those
handlers just reference global variables), so we can pass around plain function
pointers.

The previous std::function approach was also bugged as it was silently
swallowing the lambda but returning a nullptr when using
signal_handler.target<void(int)>(), as the type was mismatching.