Store a copy of the returned struct tm.
The documentation says localtime(), gmtime() and other functions will return a
pointer to a statically allocated memory area, which means the object pointed to
might be modified between calls of those functions. Let's dereference the
returned pointer so to only store a copy of the pointed structure instead.

Also fix stupidity in the test for ClockGetTimeForTimezone().

Reference:
  man 3 ctime