Add a custom Xvfb wrapper
As it turns out, xvfb-run is not universally available, and this would make
users of systems like FreeBSD unable to launch some of the test suites.

This adds a simple wrapper around Xvfb and other ancillary commands to make it
easier to run programs inside a virtual X server. If xvfb-run is available on
the system, this just shells out to it without introducing any visible
difference.