Introduce util::fs::SystemInterface and friends
This adds a layer of abstraction between util::fs and the operating
system, allowing the caller to replace the default way we interface with
the system with a custom one.

For now, this only includes stat(), as it's the only thing that we need to
override at this stage. More methods will be added as needed.

To be used in unit tests for dependency injection.

Related issues:
  - #22