fs: add Path::BaseName(), use POSIX methods
No need to roll our own code when it's already provided.

This slightly changes the behavior of Path::DirectoryName() will now return a
"." in case an empty path is given (or one without a parent directory that is
not the root).  It's fine to change this behavior as this is only currently used
by config.cc when load relative imports, which will still work.

Related issues:
  - #31