Act as theme manager when invoked as "t3"
Name not necessarily final; alternatives were either already taken by popular
software or looked unsuitable.  Suggestions welcome!

This commit only introduces the skeleton, it doesn't actually provide any real
functionality.  An install step for this is also still missing.

As the CURL dependency is going to be required for the theme manager to be able
to do any real work, this commit also flips it to required by default in the
CMakeLists.txt.

Related issues:
  - #31