execp: record if a tooltip has ever been provided
The documentation states that tooltips should not be shown if the tooltip has
been provided but empty, and will default to showing the contents of stderr if
not provided at all. This means we need an additional flag to keep track of
whether the tooltip has been provided at all.