X11: remove 'X11' and version for 'Xrandr'
Recent version of CMake ship with a modified FindX11.cmake that doesn't
work with the previous form of this find_package() invocation.
Since the X11 package should implicitly be included in this search, and
the minimum version restriction on Xrandr is on a version that's old
enough to make it unnecessary, I'm simply removing both entries here.

Fixed issues:
  - #41